height

+++ 9-binary_tree_height.c
@@ -11,8 +11,8 @@ size_t binary_tree_height(const binary_tree_t *tree)
 if (tree == NULL)
 return (0);
 
-int leftHeight = binary_tree_height(tree->left);
-int rightHeight = binary_tree_height(tree->right);
+size_t leftHeight = binary_tree_height(tree->left);
+size_t rightHeight = binary_tree_height(tree->right);
 
 if (leftHeight > rightHeight)
 return leftHeight + 1;
